Fix for line breaking algorithm
When used with sizeToFill, when measuring the one-size too big text, the previous version was leaving a line with the last character as a leftover that was never cleaned up (even when the whole word has been fitted into a previous line after decreasing font size), resulting in a last character being duplicated.

Also, in some cases, when nothing was inserted into a line during breakText, the algorithm was entering a loop that was adding an infinite number of empty lines, causing the browser to crash.